Add events component
